+++ lib/concierge/o_auth_2_client.rb
@@ -36,11 +36,11 @@ module Concierge
 
     # events that are published via +Concierge::Announcer+ and that can be listened
     # independently.
-    ON_REQUEST  = "oauth2_client.on_request"
-    ON_RESPONSE = "oauth2_client.on_response"
-    ON_FAILURE  = "oauth2_client.on_failure"
-    ON_TOKEN_REQUEST  = "oauth2_client.on_token_request"
-    ON_TOKEN_RECEIVED = "oauth2_client.on_token_received"
+    ON_REQUEST  = "oauth2_client.on_request".freeze
+    ON_RESPONSE = "oauth2_client.on_response".freeze
+    ON_FAILURE  = "oauth2_client.on_failure".freeze
+    ON_TOKEN_REQUEST = "oauth2_client.on_token_request".freeze
+    ON_TOKEN_RECEIVED = "oauth2_client.on_token_received".freeze
 
     # by default, include a self identifying +User-Agent+ HTTP header so that
     # later analysis can pinpoint the running version of Concierge (and also
@@ -48,7 +48,7 @@ module Concierge
     # ideal.)
     DEFAULT_HEADERS = {
       "User-Agent" => "Roomorama/Concierge #{Concierge::VERSION}"
-    }
+    }.freeze
 
     def initialize(id:, secret:, base_url:, token_url:, **options)
       @options = options
@@ -97,14 +97,22 @@ module Concierge
 
     private
 
-    # Fetch the access token from cache by id.
-    # TODO: - Expire the cache according to the returned token's expire_at.
-    #         Currently we request for a new token every 1 day, arbitrarily
+    # Fetch the access token from cache by id, or request a new
+    # one if it's expired already/or it wasn't even in the cache
     #
     def access_token
       return @access_token unless @access_token.nil?
 
-      token_result = cache.fetch(oauth_client.id, freshness: one_day, serializer: json_serializer ) do
+      @access_token = get_token.tap do |token|
+        if token.expired?
+          cache.invalidate(oauth_client.id)
+          get_token
+        end
+      end
+    end
+
+    def get_token
+      token_result = cache.fetch(oauth_client.id, serializer: json_serializer ) do
         token_strategy = options.fetch(:strategy, :client_credentials)
         Concierge::Announcer.trigger(ON_TOKEN_REQUEST, oauth_client, token_strategy)
         @access_token = oauth_client.public_send(token_strategy).get_token
@@ -113,10 +121,11 @@ module Concierge
       end
 
       # If cache is hit, we need to parse result into an +OAuth2::AccessToken+ object
-      @access_token = OAuth2::AccessToken.from_hash(oauth_client, token_result.value)
+      OAuth2::AccessToken.from_hash(oauth_client, token_result.value)
     end
 
     def response_with_error_handling
+      retried ||= false
       response = yield
       # No errors raised, the response is successful
       Concierge::Announcer.trigger(ON_RESPONSE, response.status, response.headers, response.body)