+++ lib/concierge/safe_access_hash.rb
@@ -49,11 +49,15 @@ module Concierge
       to_h.to_s
     end
 
+    def merge(other)
+      Concierge::SafeAccessHash.new(@hash.merge(other))
+    end
+
     # Returns array of keys that do not have value in the hash
     # Given +required_keys+ should be an array of +string+
     #
     def missing_keys_from(required_keys)
-      required_keys.select{ |k| self.get(k).nil? }
+      required_keys.select{ |k| self.get(k).to_s.empty? }
     end
   end