+++ spec/support/factories.rb
@@ -21,7 +21,13 @@ module Support
       attributes = {
         identifier: "PROP1",
         host_id:    create_host.id,
-        data:       { title: "Studio Apartment in Madrid" },
+        data:       {
+          title: "Studio Apartment in Madrid",
+          nightly_rate: 10,
+          images: [
+            { identifier: "PROP1IMAGE", url: "https://www.example.org/image.png" }
+          ]
+        },
       }.merge(overrides)
 
       property   = Property.new(attributes)