+++ CHANGELOG.md
@@ -7,6 +7,7 @@ how this file is formatted and how the process works.
 ## [Unreleased]
 ### Added
 - added `currency_code` to the list of supported attributes for booking webhook and for `API::Controllers::Params::Booking` object
+- added `get_active_properties` by Waytostay::Client, for first time import
 - add `BackgroundWorker` and related refactor on queue processing.
 - add `Workers::CalendarSynchronisation` to deal with updating the calendar of availabilities.
 - `Concierge::Cache#invalidate` and the counterpart context event.
@@ -14,8 +15,11 @@ how this file is formatted and how the process works.
 ### Changed
 - determine Roomorama API environment according to value in environment variable.
 - fixed `SafeAccessHash#get` method to return `nil` for nonexistent keys
+- fixed `SafeAccessHash#missing_keys_from` method to check for empty strings as well
 - new structure for the `config/suppliers.yml` file.
 - `Workers::Synchronisation` is now `Workers::PropertySynchronisation`.
+- OAuth2Client invalidates cache if token is expired
+- `Waytostay::Client#update_availability` to `#get_availability`, and the client use the info
 
 ### Removed
 - `update_calendar` method removed from `Roomorama::Property`.
@@ -24,6 +28,12 @@ how this file is formatted and how the process works.
 ### Added
 - Synchronisation with AtLeisure
 - Ability to enable or disable context tracking at certain points in time.
+- `Concierge::SafeAccessHash#merge` utils method
+- `waytostay#get_changes` returns changes from waytostay api
+- `waytostay#get_property` returns a roomorama property from waytostay api
+- `waytostay#update_media` sets images to all the new images from waytostay
+- `waytostay#update_availability` sets availability to all the new availability from waytostay
+- Waytostay synchronisation worker
 - `Kigo::Booking` - implemented booking with supplier
 
 ### Changed